Split //test/common/upstream:health_checker_impl_test to reduce test time
